refs #1 Create project bot
