+++ .dockerignore
@@ -1,0 +1,4 @@
+/venv
+.idea
+.git
+.dockerignore
