+++ .env
@@ -1,0 +1,7 @@
+DB_HOST = localhost
+DB_PORT = 5432
+DB_NAME = tasks
+DB_USER = manager
+DB_PASS = manager
+API_TOKEN_BOT = 1228603113:AAE_PzeNsYP3LHEyy2zHQlic-MY3QDY3tm0
+URL_SITE = https://codeforces.com/problemset?order=BY_RATING_DESC
