+++ app/bot/config.py
@@ -3,10 +3,13 @@ import os
 
 load_dotenv()
 
+API_TOKEN = os.environ.get("API_TOKEN_BOT")
+
+"""
 DB_HOST = os.environ.get("DB_HOST")
 DB_PORT = os.environ.get("DB_PORT")
 DB_NAME = os.environ.get("DB_NAME")
 DB_USER = os.environ.get("DB_USER")
 DB_PASS = os.environ.get("DB_PASS")
 API_TOKEN = os.environ.get("API_TOKEN_BOT")
-URL_SITE = os.environ.get("URL_SITE")
+URL_SITE = os.environ.get("URL_SITE")"""
