+++ app/bot/main.py
@@ -1,0 +1,25 @@
+import asyncio
+import settings
+from config import *
+from aiogram.filters.command import Command
+from aiogram import types, Bot, Dispatcher
+import logging
+
+logging.basicConfig(level=logging.INFO)
+
+bot = Bot(token=API_TOKEN)
+
+dp = Dispatcher()
+
+
+@dp.message(Command("start"))
+async def cmd_start(message: types.Message):
+    await message.answer("Hello")
+
+
+async def main():
+    await dp.start_polling(bot, skip_updates=True)
+
+
+if __name__ == "__main__":
+    asyncio.run(main())
