+++ app/bot/settings.py
@@ -1,0 +1,10 @@
+"""from app.config import API_TOKEN
+from aiogram import Bot, Dispatcher
+import logging
+
+logging.basicConfig(level=logging.INFO)
+
+bot = Bot(token=API_TOKEN)
+
+dp = Dispatcher()
+"""
