+++ docker-compose.yml
@@ -1,0 +1,11 @@
+version: "3.9"
+
+services:
+  telegram-bot:
+    container_name: tele-bot
+    working_dir: /tele_bot
+    restart: always
+    build:
+        context: .
+        dockerfile: Dockerfile
+    command: sh -c "python ./app/bot/main.py"
