--- requirements.txt
@@ -1,35 +1,0 @@
-aiofiles==23.1.0
-aiogram==3.0.0b7
-aiohttp==3.8.4
-aiosignal==1.3.1
-alembic==1.10.2
-async-timeout==4.0.2
-attrs==22.2.0
-Babel==2.12.1
-beautifulsoup4==4.11.2
-black==23.1.0
-certifi==2022.12.7
-charset-normalizer==3.1.0
-click==8.1.3
-frozenlist==1.3.3
-greenlet==2.0.2
-idna==3.4
-magic-filter==1.0.9
-Mako==1.2.4
-MarkupSafe==2.1.2
-multidict==6.0.4
-mypy-extensions==1.0.0
-packaging==23.0
-pathspec==0.11.1
-platformdirs==3.1.1
-psycopg2==2.9.5
-pydantic==1.10.6
-python-dotenv==1.0.0
-pytz==2022.7.1
-requests==2.28.2
-soupsieve==2.4
-SQLAlchemy==2.0.6
-tomli==2.0.1
-typing_extensions==4.5.0
-urllib3==1.26.15
-yarl==1.8.2
